loaded-latency: always measure TSC frequency on x86
The hwclock is used to compute bandwidth.  However, for x86,
the TSC frequency varies by system, so unless --hwclock-freq
is used to provide an accurate frequency, the reported
bandwidth is not correct.

This patch has x86 to always measure the TSC frequency instead
of using a fixed frequency.  The measurement duration is for
0.1 seconds.  This is supported by enhancing the estimation
routine for sub-second measurement.

--estimate-hwclock-freq now also uses the updated routine,
dropping the best/worst out of 12 samples.